Remove the ZCatalog implementation.
Use Solr everywhere.
