+++ src/ploneintranet/search/meta.zcml
