--- src/ploneintranet/search/tests/__init__.py
@@ -1,1 +1,0 @@
-# tests
