+++ src/ploneintranet/search/tests/test_directives.py
@@ -7,8 +7,6 @@ from ..interfaces import IConnectionConfig
 
 class TestSolrConfigDirective(testing.IntegrationTestCase):
 
-    layer = testing.INTEGRATION_TESTING
-
     def test_directive(self):
         util = queryUtility(IConnectionConfig, default=None)
         verifyObject(IConnectionConfig, util)
